requirements.txt
openai
langchain
langchain-openai
streamlit
faiss-cpu
python-dotenv
pypdf
pdfminer.six
watchdog